Changing DiceWords so it uses notifyListeners, extends ChangeNotifier, and works with provider.
This should fix the lack of re-rendering when I load a custom dice list.

Error: "Tried to use Provider with a subtype of Listenable/Stream (DiceWords)."
In main.dart, I need to change from Provider to ChangeNotifierProvider

Working. But back button doesn't load the new settings. Hav to press DONE.